move protocol buffers to same locations as generated code; start using github.com/packethost/pkg/log/logr
