+++ api/v1/bmc.proto
@@ -1,10 +1,10 @@
 syntax = "proto3";
 
-option go_package = "github.com/tinkerbell/pbnj/pkg/api/v1";
+option go_package = "github.com/tinkerbell/pbnj/api/v1";
 
-package github.com.tinkerbell.pbnj.api.proto.v1;
+package github.com.tinkerbell.pbnj.api.v1;
 
-import "api/proto/v1/common.proto";
+import "api/v1/common.proto";
 
 service BMC {
     rpc NetworkSource (NetworkSourceRequest) returns (NetworkSourceResponse);
