+++ api/v1/common.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-option go_package = "github.com/tinkerbell/pbnj/pkg/api/v1";
+option go_package = "github.com/tinkerbell/pbnj/api/v1";
 
-package github.com.tinkerbell.pbnj.api.proto.v1;
+package github.com.tinkerbell.pbnj.api.v1;
 
 message Host {
     string host = 1;
