+++ pkg/repository/repo.go
@@ -3,7 +3,7 @@
 package repository
 
 import (
-	v1 "github.com/tinkerbell/pbnj/pkg/api/v1"
+	v1 "github.com/tinkerbell/pbnj/api/v1"
 )
 
 // Actions interface for interacting with the persistence layer
