+++ server/grpcsvr/rpc/machine_test.go
@@ -6,10 +6,11 @@ import (
 
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"github.com/onsi/gomega"
+	packet_logr "github.com/packethost/pkg/log/logr"
 	"github.com/philippgille/gokv"
 	"github.com/philippgille/gokv/freecache"
+	v1 "github.com/tinkerbell/pbnj/api/v1"
 	"github.com/tinkerbell/pbnj/cmd/zaplog"
-	v1 "github.com/tinkerbell/pbnj/pkg/api/v1"
 	"github.com/tinkerbell/pbnj/server/grpcsvr/persistence"
 	"github.com/tinkerbell/pbnj/server/grpcsvr/taskrunner"
 )
@@ -54,7 +55,8 @@ func TestDevice(t *testing.T) {
 
 			ctx := context.Background()
 
-			logger, zapLogger, _ := zaplog.RegisterLogger()
+			l, zapLogger, _ := packet_logr.NewPacketLogr()
+			logger := zaplog.RegisterLogger(l)
 			ctx = ctxzap.ToContext(ctx, zapLogger)
 			f := freecache.NewStore(freecache.DefaultOptions)
 			s := gokv.Store(f)
@@ -152,7 +154,8 @@ func TestPower(t *testing.T) {
 
 			ctx := context.Background()
 
-			logger, zapLogger, _ := zaplog.RegisterLogger()
+			l, zapLogger, _ := packet_logr.NewPacketLogr()
+			logger := zaplog.RegisterLogger(l)
 			ctx = ctxzap.ToContext(ctx, zapLogger)
 			f := freecache.NewStore(freecache.DefaultOptions)
 			s := gokv.Store(f)
