+++ server/grpcsvr/rpc/task.go
@@ -3,7 +3,7 @@ package rpc
 import (
 	"context"
 
-	v1 "github.com/tinkerbell/pbnj/pkg/api/v1"
+	v1 "github.com/tinkerbell/pbnj/api/v1"
 	"github.com/tinkerbell/pbnj/pkg/logging"
 	"github.com/tinkerbell/pbnj/pkg/task"
 )
