+++ server/grpcsvr/rpc/task_test.go
@@ -7,10 +7,11 @@ import (
 
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"github.com/onsi/gomega"
+	packet_logr "github.com/packethost/pkg/log/logr"
 	"github.com/philippgille/gokv"
 	"github.com/philippgille/gokv/freecache"
+	v1 "github.com/tinkerbell/pbnj/api/v1"
 	"github.com/tinkerbell/pbnj/cmd/zaplog"
-	v1 "github.com/tinkerbell/pbnj/pkg/api/v1"
 	"github.com/tinkerbell/pbnj/pkg/oob"
 	"github.com/tinkerbell/pbnj/server/grpcsvr/persistence"
 	"github.com/tinkerbell/pbnj/server/grpcsvr/taskrunner"
@@ -26,7 +27,8 @@ func TestTaskFound(t *testing.T) {
 			Details: nil,
 		},
 	}
-	logger, zapLogger, _ := zaplog.RegisterLogger()
+	l, zapLogger, _ := packet_logr.NewPacketLogr()
+	logger := zaplog.RegisterLogger(l)
 	ctx = ctxzap.ToContext(ctx, zapLogger)
 	f := freecache.NewStore(freecache.DefaultOptions)
 	s := gokv.Store(f)
@@ -84,7 +86,8 @@ func TestRecordNotFound(t *testing.T) {
 
 			ctx := context.Background()
 
-			logger, zapLogger, _ := zaplog.RegisterLogger()
+			l, zapLogger, _ := packet_logr.NewPacketLogr()
+			logger := zaplog.RegisterLogger(l)
 			ctx = ctxzap.ToContext(ctx, zapLogger)
 			f := freecache.NewStore(freecache.DefaultOptions)
 			s := gokv.Store(f)
