+++ server/grpcsvr/rpc/user_test.go
@@ -6,8 +6,9 @@ import (
 
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"github.com/onsi/gomega"
+	packet_logr "github.com/packethost/pkg/log/logr"
+	v1 "github.com/tinkerbell/pbnj/api/v1"
 	"github.com/tinkerbell/pbnj/cmd/zaplog"
-	v1 "github.com/tinkerbell/pbnj/pkg/api/v1"
 )
 
 func TestUserCreate(t *testing.T) {
@@ -44,7 +45,8 @@ func TestUserCreate(t *testing.T) {
 
 			ctx := context.Background()
 
-			logger, zapLogger, _ := zaplog.RegisterLogger()
+			l, zapLogger, _ := packet_logr.NewPacketLogr()
+			logger := zaplog.RegisterLogger(l)
 			ctx = ctxzap.ToContext(ctx, zapLogger)
 			userSvc := UserService{
 				Log: logger,
@@ -94,7 +96,8 @@ func TestDeleteUser(t *testing.T) {
 
 			ctx := context.Background()
 
-			logger, zapLogger, _ := zaplog.RegisterLogger()
+			l, zapLogger, _ := packet_logr.NewPacketLogr()
+			logger := zaplog.RegisterLogger(l)
 			ctx = ctxzap.ToContext(ctx, zapLogger)
 			userSvc := UserService{
 				Log: logger,
@@ -147,7 +150,8 @@ func TestUpdateUser(t *testing.T) {
 
 			ctx := context.Background()
 
-			logger, zapLogger, _ := zaplog.RegisterLogger()
+			l, zapLogger, _ := packet_logr.NewPacketLogr()
+			logger := zaplog.RegisterLogger(l)
 			ctx = ctxzap.ToContext(ctx, zapLogger)
 			userSvc := UserService{
 				Log: logger,
