+++ server/grpcsvr/server_test.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
+	packet_logr "github.com/packethost/pkg/log/logr"
 	"github.com/philippgille/gokv"
 	"github.com/philippgille/gokv/freecache"
 	"github.com/tinkerbell/pbnj/cmd/zaplog"
@@ -22,7 +23,8 @@ import (
 func TestRunServer(t *testing.T) {
 	ctx := context.Background()
 	ctx, cancel := context.WithTimeout(ctx, 9*time.Second)
-	log, zapLogger, err := zaplog.RegisterLogger()
+	l, zapLogger, err := packet_logr.NewPacketLogr()
+	log := zaplog.RegisterLogger(l)
 	ctx = ctxzap.ToContext(ctx, zapLogger)
 	if err != nil {
 		t.Fatal(err)
@@ -56,7 +58,8 @@ func TestRunServerSignals(t *testing.T) {
 	ctx := context.Background()
 	ctx, cancel := context.WithTimeout(ctx, 3*time.Second)
 	defer cancel()
-	log, zapLogger, err := zaplog.RegisterLogger()
+	l, zapLogger, err := packet_logr.NewPacketLogr()
+	log := zaplog.RegisterLogger(l)
 	ctx = ctxzap.ToContext(ctx, zapLogger)
 	if err != nil {
 		t.Fatal(err)
@@ -97,7 +100,8 @@ func TestRunServerPortInUse(t *testing.T) {
 	ctx := context.Background()
 	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
 	defer cancel()
-	log, zapLogger, err := zaplog.RegisterLogger()
+	l, zapLogger, err := packet_logr.NewPacketLogr()
+	log := zaplog.RegisterLogger(l)
 	ctx = ctxzap.ToContext(ctx, zapLogger)
 	if err != nil {
 		t.Fatal(err)
